be evil: parse json wity python
'git annex status --fast' supports output in JSON with '--json' switch. However
zsh does not have a builtin json parser :( So i added a python snippet to parse
the output. This should be somewhat portable, since Python is in many
distributions and the 'json' module is in the standar distribution, but
obvioulsy problematic if I ever attempt to send the completion upstream. It's
unlikely that anyone will ever include python as a dependency of zsh...

Also, this solution crashes badly, if we attempt to invoke the completion
outside of an annex.