be evil again: for repository completion
Again use python to parse the json. This time its even worse, because we call
'git annex status --fast --json' three times. Maybe something with 'eval' can be
done here, but then again that's evil too.

Also, not sure if the description will work for all the options.... Some, for
example '--from' say it needs to be a configured remote.